Move GUI Rendering Code into separate files
