+++ atcg_lib/include/Scene/Components/EdgeRenderComponent.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <Scene/Components/RenderComponent.h>
+#include <Scene/ComponentGUIHandler.h>
 
 namespace atcg
 {
@@ -13,4 +14,10 @@ struct EdgeRenderComponent : public RenderComponent
     glm::vec3 color = glm::vec3(1);
     // TODO Edge radius?
 };
+
+namespace GUI
+{
+ATCG_DECLARE_COMPONENT_GUI_RENDERER(EdgeRenderComponent);
+}
+
 }    // namespace atcg
