+++ atcg_lib/include/Scene/Components/InstanceRenderComponent.h
@@ -2,6 +2,7 @@
 
 #include <Asset/AssetManagerSystem.h>
 #include <Scene/Components/RenderComponent.h>
+#include <Scene/ComponentGUIHandler.h>
 
 namespace atcg
 {
@@ -32,4 +33,10 @@ struct InstanceRenderComponent : public RenderComponent
     AssetHandle material_handle = 0;
     AssetHandle shader_handle;
 };
+
+namespace GUI
+{
+ATCG_DECLARE_COMPONENT_GUI_RENDERER(InstanceRenderComponent);
+}
+
 }    // namespace atcg
