+++ atcg_lib/include/Scene/Components/PointSphereRenderComponent.h
@@ -5,6 +5,7 @@
 #include <Renderer/Shader.h>
 #include <Renderer/ShaderManager.h>
 #include <Scene/Components/RenderComponent.h>
+#include <Scene/ComponentGUIHandler.h>
 
 namespace atcg
 {
@@ -46,4 +47,10 @@ struct PointSphereRenderComponent : public RenderComponent
     AssetHandle material_handle = 0;
     AssetHandle shader_handle   = 0;
 };
+
+namespace GUI
+{
+ATCG_DECLARE_COMPONENT_GUI_RENDERER(PointSphereRenderComponent);
+}
+
 }    // namespace atcg
