+++ atcg_lib/platform/glfw/src/Scene/ComponentGUIHandler.cpp
@@ -12,578 +12,6 @@ namespace atcg
 namespace GUI
 {
 
-bool displayTransform(const std::string& id, TransformComponent& transform)
-{
-    bool updated       = false;
-    glm::vec3 position = transform.getPosition();
-    std::stringstream label;
-    label << "Position##" << id;
-    if(ImGui::DragFloat3(label.str().c_str(), glm::value_ptr(position), 0.05f))
-    {
-        transform.setPosition(position);
-        updated = true;
-    }
-    glm::vec3 scale = transform.getScale();
-    label.str(std::string());
-    label << "Scale##" << id;
-    if(ImGui::DragFloat3(label.str().c_str(), glm::value_ptr(scale), 0.05f, 1e-5f, FLT_MAX))
-    {
-        scale = glm::clamp(scale, 1e-5f, FLT_MAX);
-        transform.setScale(scale);
-        updated = true;
-    }
-    glm::vec3 rotation = glm::degrees(transform.getRotation());
-    label.str(std::string());
-    label << "Rotation##" << id;
-    if(ImGui::DragFloat3(label.str().c_str(), glm::value_ptr(rotation), 0.05f))
-    {
-        transform.setRotation(glm::radians(rotation));
-        updated = true;
-    }
-
-    return updated;
-}
-
-void ComponentGUIRenderer<TransformComponent>::draw_component(const atcg::ref_ptr<Scene>& scene,
-                                                              Entity entity,
-                                                              TransformComponent& transform) const
-{
-    std::string id = std::to_string(entity.getComponent<IDComponent>().ID());
-
-    TransformComponent transform_ = transform;
-
-    bool updated = displayTransform(id, transform_);
-
-    if(updated)
-    {
-        RevisionStack::startRecording<ComponentEditedRevision<TransformComponent>>(scene, entity);
-        transform = transform_;
-        atcg::RevisionStack::endRecording();
-    }
-}
-
-
-void ComponentGUIRenderer<CameraComponent>::draw_component(const atcg::ref_ptr<Scene>& scene,
-                                                           Entity entity,
-                                                           CameraComponent& _component) const
-{
-    CameraComponent component = _component;
-    bool updated              = false;
-
-    float content_scale = atcg::Application::get()->getWindow()->getContentScale();
-    std::string id      = std::to_string(entity.getComponent<IDComponent>().ID());
-
-    atcg::ref_ptr<atcg::PerspectiveCamera> camera =
-        std::dynamic_pointer_cast<atcg::PerspectiveCamera>(component.camera->copy());
-
-    atcg::CameraIntrinsics intrinsics = camera->getIntrinsics();
-    glm::mat3 K = atcg::CameraUtils::convert_to_opencv(intrinsics, component.width, component.height);
-
-    float fx = K[0][0];
-    float fy = K[1][1];
-    float cx = K[0][2];
-    float cy = K[1][2];
-
-    float f[2]      = {fx, fy};
-    float c[2]      = {cx, cy};
-    uint32_t res[2] = {component.width, component.height};
-    float offset[2] = {intrinsics.opticalCenter().x, intrinsics.opticalCenter().y};
-
-    float aspect_ratio = intrinsics.aspectRatio();
-    float fov          = intrinsics.FOV();
-
-    std::stringstream label;
-    label << "Aspect Ratio##" << id;
-    if(ImGui::DragFloat(label.str().c_str(), &aspect_ratio, 0.05f, 0.1f, 5.0f))
-    {
-        intrinsics.setAspectRatio(aspect_ratio);
-        updated = true;
-    }
-
-    label.str(std::string());
-    label << "FOV##" << id;
-    if(ImGui::DragFloat(label.str().c_str(), &fov, 0.5f, 10.0f, 120.0f))
-    {
-        intrinsics.setFOV(fov);
-        updated = true;
-    }
-
-    label.str(std::string());
-    label << "Resolution##" << id;
-    if(ImGui::DragInt2(label.str().c_str(), (int*)res, 1, 1, 4096))
-    {
-        component.width  = res[0];
-        component.height = res[1];
-        updated          = true;
-    }
-
-    label.str(std::string());
-    label << "Optical Center##" << id;
-    if(ImGui::DragFloat2(label.str().c_str(), offset, 0.01f, -1.0f, 1.0f))
-    {
-        intrinsics.setOpticalCenter(glm::make_vec2(offset));
-        updated = true;
-    }
-
-    ImGui::Separator();
-
-    label.str(std::string());
-    label << "Focal Length##" << id;
-    if(ImGui::DragFloat2(label.str().c_str(), f, 0.5f, 1.0f, 4096.0f))
-    {
-        intrinsics = atcg::CameraUtils::convert_from_opencv(f[0],
-                                                            f[1],
-                                                            c[0],
-                                                            c[1],
-                                                            intrinsics.zNear(),
-                                                            intrinsics.zFar(),
-                                                            component.width,
-                                                            component.height);
-        updated    = true;
-    }
-
-    label.str(std::string());
-    label << "Principal Point##" << id;
-    if(ImGui::DragFloat2(label.str().c_str(), c, 0.5f, 1.0f, 4096.0f))
-    {
-        intrinsics = atcg::CameraUtils::convert_from_opencv(f[0],
-                                                            f[1],
-                                                            c[0],
-                                                            c[1],
-                                                            intrinsics.zNear(),
-                                                            intrinsics.zFar(),
-                                                            component.width,
-                                                            component.height);
-        updated    = true;
-    }
-
-    label.str(std::string());
-    label << "Color##" << id;
-    updated = ImGui::ColorEdit3(label.str().c_str(), glm::value_ptr(component.color)) || updated;
-
-    label.str(std::string());
-    label << "Scale##" << id;
-    updated = ImGui::DragFloat(label.str().c_str(), &component.render_scale, 0.01f, 0.01f, FLT_MAX) || updated;
-
-    uint32_t preview_height = 128;
-    uint32_t preview_width =
-        glm::clamp((uint32_t)(float(component.width) / float(component.height) * 128.0f), uint32_t(1), uint32_t(4096));
-
-    if(!component.preview || component.preview->width() != preview_width ||
-       component.preview->height() != preview_height)
-    {
-        component.preview = atcg::make_ref<atcg::Framebuffer>(preview_width, preview_height);
-        component.preview->attachColor();
-        component.preview->attachDepth();
-        component.preview->complete();
-
-        _component.preview = component.preview;    // This should not count as an update, this is just a lazy init
-        // updated = true;
-    }
-
-    atcg::Dictionary context;
-    context.setValue("camera", component.camera);
-    context.setValue("target", component.preview);
-    context.setValue("draw_cameras", false);
-    scene->draw(context);
-
-    updated = ImGui::Checkbox("Show Preview##cam", &component.render_preview) || updated;
-
-    uint64_t textureID = component.preview->getColorAttachement(0)->getID();
-
-    ImVec2 window_size = ImGui::GetWindowSize();
-    ImGui::SetCursorPos(ImVec2((window_size.x - preview_width) * 0.5f, ImGui::GetCursorPosY()));
-    ImGui::Image((ImTextureID)textureID,
-                 ImVec2(content_scale * preview_width, content_scale * preview_height),
-                 ImVec2 {0, 1},
-                 ImVec2 {1, 0});
-
-    if(ImGui::Button("Screenshot"))
-    {
-        auto t  = std::time(nullptr);
-        auto tm = *std::localtime(&t);
-        std::ostringstream oss;
-
-        std::string_view tag = entity.getComponent<NameComponent>().name();
-
-        oss << "bin/" << tag << "_" << std::put_time(&tm, "%Y-%m-%d_%H-%M-%S") << ".png";
-
-        atcg::Utils::screenshot(scene, component.camera, component.width, component.height, oss.str());
-    }
-
-    ImGui::Separator();
-
-    auto new_handle = displayTexture2DSelection("camera", component.image_handle);
-
-    updated                = (new_handle != component.image_handle) || updated;
-    component.image_handle = new_handle;
-
-    ImGui::Separator();
-
-    auto scene_camera = scene->getCamera();
-
-    if(scene_camera)
-    {
-        if(ImGui::Button("Fly to"))
-        {
-            scene_camera->setExtrinsics(component.camera->getExtrinsics());
-        }
-
-        if(ImGui::Button("Set from View"))
-        {
-            camera->setExtrinsics(scene_camera->getExtrinsics());
-            intrinsics = scene_camera->getIntrinsics();
-
-            component.width  = atcg::Renderer::getFramebuffer()->width();
-            component.height = atcg::Renderer::getFramebuffer()->height();
-
-            if(entity.hasComponent<atcg::TransformComponent>())
-            {
-                entity.getComponent<atcg::TransformComponent>().setModel(glm::inverse(scene_camera->getView()));
-            }
-
-            updated = true;
-        }
-    }
-
-    if(updated)
-    {
-        ATCG_DEBUG("Updated");
-        atcg::RevisionStack::startRecording<ComponentEditedRevision<CameraComponent>>(scene, entity);
-        _component = component;
-        camera->setIntrinsics(intrinsics);
-        _component.camera = camera;
-        atcg::RevisionStack::endRecording();
-    }
-
-    if(entity.hasComponent<atcg::TransformComponent>())
-    {
-        glm::mat4 model = entity.getComponent<atcg::TransformComponent>().getModel();
-
-        float scale_x = glm::length(glm::vec3(model[0]));
-        float scale_y = glm::length(glm::vec3(model[1]));
-        float scale_z = glm::length(glm::vec3(model[2]));
-
-        model = model * glm::scale(glm::vec3(1.0f / scale_x, 1.0f / scale_y, 1.0f / scale_z));
-
-        camera->setView(glm::inverse(model));
-        _component.camera = camera;
-    }
-}
-
-
-void ComponentGUIRenderer<GeometryComponent>::draw_component(const atcg::ref_ptr<Scene>& scene,
-                                                             Entity entity,
-                                                             GeometryComponent& component) const
-{
-    auto new_handle = displayGraphSelection("geometry", component.graph_handle);
-    bool updated    = (new_handle != component.graph_handle);
-
-    if(updated)
-    {
-        RevisionStack::startRecording<ComponentEditedRevision<GeometryComponent>>(scene, entity);
-        component.graph_handle = new_handle;
-        atcg::RevisionStack::endRecording();
-    }
-}
-
-
-void ComponentGUIRenderer<MeshRenderComponent>::draw_component(const atcg::ref_ptr<Scene>& scene,
-                                                               Entity entity,
-                                                               MeshRenderComponent& component) const
-{
-    MeshRenderComponent component_copy = component;
-    bool updated                       = ImGui::Checkbox("Visible##visiblemesh", &component_copy.visible);
-
-    // Material
-    auto material_handle = component_copy.material_handle;
-    auto shader_handle   = component_copy.shader_handle;
-
-    auto new_handle                = displayMaterialSelection("mesh", material_handle);
-    updated                        = (new_handle != material_handle) || updated;
-    component_copy.material_handle = new_handle;
-
-    new_handle                   = displayShaderSelection("mesh", shader_handle);
-    updated                      = (new_handle != shader_handle) || updated;
-    component_copy.shader_handle = new_handle;
-
-    updated = ImGui::Checkbox("Receive Shadows##MeshRenderComponent", &component_copy.receive_shadow) || updated;
-
-    if(updated)
-    {
-        atcg::RevisionStack::startRecording<ComponentEditedRevision<MeshRenderComponent>>(scene, entity);
-        component = component_copy;
-        atcg::RevisionStack::endRecording();
-    }
-}
-
-
-void ComponentGUIRenderer<PointRenderComponent>::draw_component(const atcg::ref_ptr<Scene>& scene,
-                                                                Entity entity,
-                                                                PointRenderComponent& _component) const
-{
-    std::string id = std::to_string(entity.getComponent<IDComponent>().ID());
-
-    PointRenderComponent component = _component;
-
-    bool updated    = ImGui::Checkbox("Visible##visiblepoints", &component.visible);
-    glm::vec3 color = component.color;
-    std::stringstream label;
-    label << "Base Color##point" << id;
-    if(ImGui::ColorEdit3(label.str().c_str(), glm::value_ptr(color)))
-    {
-        component.color = color;
-        updated         = true;
-    }
-
-    int point_size = (int)component.point_size;
-    label.str(std::string());
-    label << "Point Size##point" << id;
-    if(ImGui::DragInt(label.str().c_str(), &point_size, 1, 1, INT_MAX))
-    {
-        component.point_size = (float)point_size;
-        updated              = true;
-    }
-
-    auto shader_handle      = component.shader_handle;
-    auto new_handle         = displayShaderSelection("point", shader_handle);
-    updated                 = (new_handle != shader_handle) || updated;
-    component.shader_handle = new_handle;
-
-    if(updated)
-    {
-        atcg::RevisionStack::startRecording<ComponentEditedRevision<PointRenderComponent>>(scene, entity);
-        _component = component;
-        atcg::RevisionStack::endRecording();
-    }
-}
-
-
-void ComponentGUIRenderer<PointSphereRenderComponent>::draw_component(const atcg::ref_ptr<Scene>& scene,
-                                                                      Entity entity,
-                                                                      PointSphereRenderComponent& _component) const
-{
-    PointSphereRenderComponent component = _component;
-    std::string id                       = std::to_string(entity.getComponent<IDComponent>().ID());
-
-    bool updated = ImGui::Checkbox("Visible##visiblepointsphere", &component.visible);
-
-    float point_size = component.point_size;
-    std::stringstream label;
-    label << "Point Size##pointsphere" << id;
-    if(ImGui::DragFloat(label.str().c_str(), &point_size, 0.001f, 0.001f, FLT_MAX / INT_MAX))
-    {
-        component.point_size = point_size;
-        updated              = true;
-    }
-
-    // Material
-    auto material_handle = component.material_handle;
-
-    auto new_handle           = displayMaterialSelection("pointsphere", material_handle);
-    updated                   = (new_handle != material_handle) || updated;
-    component.material_handle = new_handle;
-
-    auto shader_handle      = component.shader_handle;
-    new_handle              = displayShaderSelection("pointsphere", shader_handle);
-    updated                 = (new_handle != shader_handle) || updated;
-    component.shader_handle = new_handle;
-
-    if(updated)
-    {
-        atcg::RevisionStack::startRecording<ComponentEditedRevision<PointSphereRenderComponent>>(scene, entity);
-        _component = component;
-        atcg::RevisionStack::endRecording();
-    }
-}
-
-
-void ComponentGUIRenderer<EdgeRenderComponent>::draw_component(const atcg::ref_ptr<Scene>& scene,
-                                                               Entity entity,
-                                                               EdgeRenderComponent& _component) const
-{
-    EdgeRenderComponent component = _component;
-
-    std::string id = std::to_string(entity.getComponent<IDComponent>().ID());
-
-    bool updated    = ImGui::Checkbox("Visible##visibleedge", &component.visible);
-    glm::vec3 color = component.color;
-    std::stringstream label;
-    label << "Base Color##edge" << id;
-    if(ImGui::ColorEdit3(label.str().c_str(), glm::value_ptr(color)))
-    {
-        component.color = color;
-        updated         = true;
-    }
-
-    if(updated)
-    {
-        atcg::RevisionStack::startRecording<ComponentEditedRevision<EdgeRenderComponent>>(scene, entity);
-        _component = component;
-        atcg::RevisionStack::endRecording();
-    }
-}
-
-
-void ComponentGUIRenderer<EdgeCylinderRenderComponent>::draw_component(const atcg::ref_ptr<Scene>& scene,
-                                                                       Entity entity,
-                                                                       EdgeCylinderRenderComponent& _component) const
-{
-    EdgeCylinderRenderComponent component = _component;
-    std::string id                        = std::to_string(entity.getComponent<IDComponent>().ID());
-
-    bool updated = ImGui::Checkbox("Visible##visibleedgecylinder", &component.visible);
-    std::stringstream label;
-    label << "Radius##edgecylinder" << id;
-    float radius = component.radius;
-    if(ImGui::DragFloat(label.str().c_str(), &radius, 0.001f, 0.001f, FLT_MAX / INT_MAX))
-    {
-        component.radius = radius;
-        updated          = true;
-    }
-
-    // Material
-    auto material_handle = component.material_handle;
-
-    auto new_handle           = displayMaterialSelection("edgecylinder", material_handle);
-    updated                   = (new_handle != material_handle) || updated;
-    component.material_handle = new_handle;
-
-    if(updated)
-    {
-        atcg::RevisionStack::startRecording<ComponentEditedRevision<EdgeCylinderRenderComponent>>(scene, entity);
-        _component = component;
-        atcg::RevisionStack::endRecording();
-    }
-}
-
-
-void ComponentGUIRenderer<InstanceRenderComponent>::draw_component(const atcg::ref_ptr<Scene>& scene,
-                                                                   Entity entity,
-                                                                   InstanceRenderComponent& _component) const
-{
-    InstanceRenderComponent component = _component;
-    bool updated                      = ImGui::Checkbox("Visible##visibleinstance", &component.visible);
-
-    // Material
-    auto material_handle = component.material_handle;
-
-    auto new_handle           = displayMaterialSelection("instance", material_handle);
-    updated                   = (new_handle != material_handle) || updated;
-    component.material_handle = new_handle;
-    updated = ImGui::Checkbox("Receive Shadows##InstanceRenderComponent", &component.receive_shadow) || updated;
-
-    auto shader_handle      = component.shader_handle;
-    new_handle              = displayShaderSelection("instance", shader_handle);
-    updated                 = (new_handle != shader_handle) || updated;
-    component.shader_handle = new_handle;
-
-    if(updated)
-    {
-        atcg::RevisionStack::startRecording<ComponentEditedRevision<InstanceRenderComponent>>(scene, entity);
-        _component = component;
-        atcg::RevisionStack::endRecording();
-    }
-}
-
-
-void ComponentGUIRenderer<PointLightComponent>::draw_component(const atcg::ref_ptr<Scene>& scene,
-                                                               Entity entity,
-                                                               PointLightComponent& _component) const
-{
-    PointLightComponent component = _component;
-    bool updated = ImGui::DragFloat("Intensity##PointLight", &component.intensity, 0.01f, 0.0f, FLT_MAX);
-    updated      = ImGui::ColorEdit3("Color##PointLight", glm::value_ptr(component.color)) || updated;
-    updated      = ImGui::Checkbox("Cast Shadows##PointLight", &component.cast_shadow) || updated;
-
-    if(updated)
-    {
-        atcg::RevisionStack::startRecording<ComponentEditedRevision<PointLightComponent>>(scene, entity);
-        _component = component;
-        atcg::RevisionStack::endRecording();
-    }
-}
-
-void ComponentGUIRenderer<MeshLightComponent>::draw_component(const atcg::ref_ptr<Scene>& scene,
-                                                              Entity entity,
-                                                              MeshLightComponent& _component) const
-{
-    MeshLightComponent component = _component;
-
-    float updated = false;
-    {
-        // auto spec        = component.getEmissiveTexture()->getSpecification();
-        // bool useTextures = spec.width != 1 || spec.height != 1;
-
-        updated = ImGui::DragFloat("Scaling", &component.intensity, 0.005f, 0.0f, FLT_MAX) || updated;
-
-        if(!AssetManager::isAssetHandleValid(component.emissive_handle))
-        {
-            auto emissive = component.getEmissiveTexture()->getData(atcg::CPU);
-
-            float color[4] = {emissive.index({0, 0, 0}).item<float>() / 255.0f,
-                              emissive.index({0, 0, 1}).item<float>() / 255.0f,
-                              emissive.index({0, 0, 2}).item<float>() / 255.0f,
-                              emissive.index({0, 0, 3}).item<float>() / 255.0f};
-
-            if(ImGui::ColorEdit4("Emissive##mesh_light", color))
-            {
-                glm::vec4 new_color = glm::make_vec4(color);
-                component.setEmissiveColor(new_color);
-                updated = true;
-            }
-        }
-
-        ImGui::Separator();
-
-        auto new_handle = displayTexture2DSelection("meshlight", component.emissive_handle);
-
-        updated                   = (new_handle != component.emissive_handle) || updated;
-        component.emissive_handle = new_handle;
-
-        ImGui::Separator();
-    }
-
-    if(updated)
-    {
-        atcg::RevisionStack::startRecording<ComponentEditedRevision<MeshLightComponent>>(scene, entity);
-        _component = component;
-        atcg::RevisionStack::endRecording();
-    }
-}
-
-
-void ComponentGUIRenderer<ScriptComponent>::draw_component(const atcg::ref_ptr<Scene>& scene,
-                                                           Entity entity,
-                                                           ScriptComponent& _component) const
-{
-    auto new_handle = displayScriptSelection("script", _component.script_handle);
-    bool updated    = (new_handle != _component.script_handle);
-
-    if(updated)
-    {
-        RevisionStack::startRecording<ComponentEditedRevision<ScriptComponent>>(scene, entity);
-        if(_component.script())
-        {
-            auto behavior = _component.behavior(scene, entity);
-            if(behavior)
-            {
-                behavior->onDetach();
-            }
-        }
-        _component.script_handle = new_handle;
-        if(_component.script())
-        {
-            auto behavior = _component.behavior(scene, entity, true);
-            if(behavior)
-            {
-                behavior->onAttach();
-            }
-        }
-        atcg::RevisionStack::endRecording();
-    }
-}
-
 AssetHandle displayMaterialSelection(const std::string& key, AssetHandle handle)
 {
     const auto& data = AssetManager::getMetaData(handle);
@@ -885,79 +313,6 @@ AssetHandle displayTexture3DSelection(const std::string& key, AssetHandle handle
     }
 
     return current_item;
-}
-
-void ComponentGUIRenderer<HomogeneousMediumComponent>::draw_component(const atcg::ref_ptr<Scene>& scene,
-                                                                      Entity entity,
-                                                                      HomogeneousMediumComponent& component) const
-{
-#ifndef ATCG_HEADLESS
-    HomogeneousMediumComponent _component = component;
-
-    bool updated = false;
-    updated      = ImGui::DragFloat("Density##homogen", &_component.density, 0.05f, 0.0f, 50.0f) || updated;
-    updated      = ImGui::DragFloat("g##homogen", &_component.g, 0.01f, -1.0f, 1.0f) || updated;
-    updated      = ImGui::ColorEdit3("albedo##homogen", glm::value_ptr(_component.albedo)) || updated;
-    updated      = ImGui::DragFloat("Le##homogen", &_component.Le, 0.01f, 0.0f, 100.0f) || updated;
-    updated      = ImGui::ColorEdit3("LeColor##homogen", glm::value_ptr(_component.Le_color)) || updated;
-
-    if(updated)
-    {
-        atcg::RevisionStack::startRecording<ComponentEditedRevision<HomogeneousMediumComponent>>(scene, entity);
-        component = _component;
-        atcg::RevisionStack::endRecording();
-    }
-#endif
-}
-
-void ComponentGUIRenderer<HeterogeneousMediumComponent>::draw_component(const atcg::ref_ptr<Scene>& scene,
-                                                                        Entity entity,
-                                                                        HeterogeneousMediumComponent& component) const
-{
-#ifndef ATCG_HEADLESS
-    HeterogeneousMediumComponent _component = component;
-
-    bool updated = false;
-
-    ImGui::Text("Density");
-    auto new_handle                = displayTexture3DSelection("densitytexture3d", _component.density_grid.handle);
-    updated                        = updated || (new_handle != _component.density_grid.handle);
-    _component.density_grid.handle = new_handle;
-    updated =
-        ImGui::DragFloat("Desity Scale##texture3d", &_component.density_grid.scale, 0.01f, 0.0f, 10.0f) || updated;
-    ImGui::Text("Bounding Box");
-    updated = ImGui::DragFloat3("Min##density", glm::value_ptr(_component.density_grid.bbox.min), 0.05f) || updated;
-    updated = ImGui::DragFloat3("Max##density", glm::value_ptr(_component.density_grid.bbox.max), 0.05f) || updated;
-
-    ImGui::Separator();
-    ImGui::Text("Albedo");
-    new_handle                    = displayTexture3DSelection("albedotexture3d", _component.albedo_grid.handle);
-    updated                       = updated || (new_handle != _component.albedo_grid.handle);
-    _component.albedo_grid.handle = new_handle;
-    updated = ImGui::DragFloat("Albedo Scale##texture3d", &_component.albedo_grid.scale, 0.01f, 0.0f, 1.0f) || updated;
-    ImGui::Text("Bounding Box");
-    updated = ImGui::DragFloat3("Min##albedo", glm::value_ptr(_component.albedo_grid.bbox.min), 0.05f) || updated;
-    updated = ImGui::DragFloat3("Max##albedo", glm::value_ptr(_component.albedo_grid.bbox.max), 0.05f) || updated;
-
-    ImGui::Separator();
-    ImGui::Text("Emission");
-    new_handle                      = displayTexture3DSelection("emissiontexture3d", _component.emission_grid.handle);
-    updated                         = updated || (new_handle != _component.emission_grid.handle);
-    _component.emission_grid.handle = new_handle;
-    updated =
-        ImGui::DragFloat("Emission Scale##texture3d", &_component.emission_grid.scale, 0.01f, 0.0f, 10.0f) || updated;
-    ImGui::Text("Bounding Box");
-    updated = ImGui::DragFloat3("Min##emission", glm::value_ptr(_component.emission_grid.bbox.min), 0.05f) || updated;
-    updated = ImGui::DragFloat3("Max##emission", glm::value_ptr(_component.emission_grid.bbox.max), 0.05f) || updated;
-    updated = ImGui::DragFloat("g##het", &_component.g, 0.01f, -1.0f, 1.0f) || updated;
-
-    if(updated)
-    {
-        atcg::RevisionStack::startRecording<ComponentEditedRevision<HeterogeneousMediumComponent>>(scene, entity);
-        component = _component;
-        atcg::RevisionStack::endRecording();
-    }
-#endif
 }
 
 }    // namespace GUI
