+++ atcg_lib/src/Scene/Components/PointLightComponent.cpp
@@ -3,5 +3,27 @@
 
 namespace atcg
 {
+namespace GUI
+{
+void ComponentGUIRenderer<PointLightComponent>::draw_component(const atcg::ref_ptr<Scene>& scene,
+                                                               Entity entity,
+                                                               PointLightComponent& _component) const
+{
+#ifndef ATCG_HEADLESS
+    PointLightComponent component = _component;
+    bool updated = ImGui::DragFloat("Intensity##PointLight", &component.intensity, 0.01f, 0.0f, FLT_MAX);
+    updated      = ImGui::ColorEdit3("Color##PointLight", glm::value_ptr(component.color)) || updated;
+    updated      = ImGui::Checkbox("Cast Shadows##PointLight", &component.cast_shadow) || updated;
+
+    if(updated)
+    {
+        atcg::RevisionStack::startRecording<ComponentEditedRevision<PointLightComponent>>(scene, entity);
+        _component = component;
+        atcg::RevisionStack::endRecording();
+    }
+#endif
+}
+}    // namespace GUI
+
 ATCG_REGISTER_COMPONENT(PointLightComponent);
 }    // namespace atcg
